созадана фабрика вопросов
